requirements.txt
pytest>=7.0.0
GitPython>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
